Move preliminary API into the registrar bundle
Instead of having the API in a seperate bundle (that will be provided by
the OSGi spec) this embeeds the preliminary API in the registrar and
allow it to be substituted.

This makes handling of this in the transition phase much easier and do
not lead to two bundles that later export the exact same package version
(but probably slightly different content).

Also the readme is added and the code is relicensed under EPL 2.0